refactor: Types and formatters
